tested "facing"
I'm not sure if it worked without the f, but I also tested it using Debug.Log();

This just prints to the console in unity. I have commented it out for now, but I can probably just be removed.

"facing" determines which way the character is looking. If the player is not looking up or down, they will just be looking left or right. If they are looking up or down actively, this will keep track of it.